Co-locate assertInt32 and related functions
